Added aliaser
Necessary to define custom aliases for default styles. This is expecially useful when combining multiple styles as a single alias.

This object works as a wrapper around the defaults where if an alias doesn't exist, it will fall back to the default. Only when nothing is found will a failure occur.