Added decoder
Necessary to decode a color encoded string into a hash representation that can be manipulated further (also handy for testing purposes).

The Reek issues are a concern but there isn't a way to simplify the logic of the `#scan` method at risk of reducing readability so am excluding these issues for now.